Fix webservice API
Previously, the webservice used a literal web browser interface, which
accepted, as text, the list of tuples from the `data` folder, and then
displayed scores on a subsequent page.

However, this is not consistent with RESTful principles.

Specifically, that webservice did not allow for `POST`ing of data, and
so did not return the scores as a response to this `POST`.

Accordingly, using the `flask_restful` library, define a RESTful
webservice that accepts JSON strings representing the parsed Amazon
Products API responses, then invokes the `MLEngine` to calculate
`scores`, then returns those `scores` to the sender of the `POST`.

Also, this commit includes `JSONtuple.py` file, which contains functions
for converting between a List of Tuples and its JSON representation.