Make engine stateless
Previously, the engine took in query information and performed the text
similarity calculation during its instantiation.

However, this requires storing and referring to object attributes, which
implicitly generates "states" in the engine.

This is undesirable, since it is conceivable that the text similarity
engine could be receiving numerous simultaneous requests, and in such a
scenario the state of an individual text similarity engine could be
disturbed, leading to undetermined behavior (and undetermined results).

Accordingly, refactor the engine such that its text similarity
calculations are performed statelessly; no information is stored in
object attributes, all are passed through the object's methods.