Pass and display human-friendly errors
Previously, assertion statements were used to validate input that was
passed to the `TextSimilarityEngine`.

However, although the `TextSimilarityEngine` could raise these
exceptions to its invoker, the `REST-webservice`, the webservice was
unable to pass these exceptions to *its* invoker, the
`TestToScores_REST` client.

As such, although information was available for parsing and display on
the server-side, that information was not being passed onto the client.

This led to the client not knowing what caused the exceptions.

Accordingly, the solution is to use a data structure that can be passed
to the client to represent the error incurred.

A `dict` seems ideally-suited for this purpose, since it is easily
JSON-serializable.

The webservice is thus able to pass on whatever object it receives from
the `TextSimilarityEngine`, whether that is a `list` (indicating that
scores were successfully calculated) or a `dict` (so far, indicating
that an error occurred in the `validateInput()` method).

The client that invoked the webservice is then able to parse the
response received (currently, whether a `list` or a `dict`) and to
display the scores, or the error message, as the case may be.