Fix signature in JSON parser
Previously, `JSONparser` accepted a `listOfTuples` and
returned a `json` object, or vice versa.

Firstly, `listOfTuples` has been split into a `query` and
a `documents` component, which it always contained.
This is in order to make the various functions that deal with
these values more explicit and intelligible.

Secondly, since we are dealing with nested `json` objects,
it is necessary for us to convert the nested `json` into a
string before it is passed to the webservice via `requests.post()`.

Since we can expect that the webservice will be receiving a
string, and that the client will be sending a string, we can
fold this into the `JSONparser` so that we don't have to
do more conversion than is necessary in the webservice and
client.

Accordingly, this file now accepts the updated (query, documents)
signature, and additionally performs conversion from that
(query, documents) pair to the string representation of a `json`
object, and vice versa.