Fix signature in webservice and client
In short, the webservice and client now use the
(query, documents) signature pair rather than the previous
`listOfTuples` conglomerate.

Additionally, neither the webservice nor the client *need* to
do string-to-json parsing or vice versa, although the webservice
still parses the json string so that it can be pretty printed
as debugging output.

For a more detailed discussion of the changes made and their
rationale, see the previous commit on this branch.