Update format of text files
Previously, these text files were stored in the `listOfTuples` format.

However, the signatures of the client, the webservice, and the
`TextSimilarityEngine` have been changed to accept the (query,
documents) signature pair.

Accordingly, update the text files so that the query is the string on
the first line, and the documents are represented by the list of tuples
on the second line.